Change expression/address values to signed, use stdint types

+++ Archs/Architecture.cpp
@@ -41,7 +41,7 @@ void ArchitectureCommand::writeTempData(TempData& tempData) const
 void ArchitectureCommand::writeSymData(SymbolData& symData) const
 {
 	// TODO: find a less ugly way to check for undefined memory positions
-	if (position == (u64)-1)
+	if (position == -1)
 		return;
 	
 	if (symText.size() != 0)
